requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
argcomplete==3.0.2
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
biopython==1.81
boto==2.49.0
boto3==1.26.89
botocore==1.29.89
cachetools==5.3.0
# certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==3.0.1
cloudpickle==2.2.1
clr-loader==0.2.5
crcmod==1.7
cryptography==39.0.2
cycler==0.11.0
fasteners==0.18
filelock==3.9.0
flatbuffers==23.3.3
fonttools==4.38.0
frozenlist==1.3.3
gast==0.4.0
gcs-oauth2-boto-plugin==3.0
google-apitools==0.5.32
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
google-reauth==0.1.1
grpcio==1.53.0
gsutil==5.21
gym==0.26.2
gym-notices==0.0.8
h5py==3.8.0
httplib2==0.20.4
huggingface-hub==0.12.1
idna==3.4
importlib-metadata==5.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
learn2learn==0.1.7
libclang==16.0.0
lmdb==1.4.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
monotonic==1.6
multidict==6.0.4
numpy==1.21.6
oauth2client==4.1.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyg-lib==0.1.0+pt113cu116
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-version==0.0.2
pythonnet==3.0.1
pytz==2022.7.1
pyu2f==0.1.5
PyYAML==6.0
qpth==0.0.15
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
retry-decorator==1.1.1
rsa==4.7.2
s3transfer==0.6.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0.post1
tape-proteins==0.5
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1+cu116
torch-cluster==1.6.0+pt113cu116
torch-geometric==2.2.0
torch-scatter==2.1.0+pt113cu116
torch-sparse==0.6.16+pt113cu116
torch-spline-conv==1.2.1+pt113cu116
torchaudio==0.13.1+cu116
torchsummary==1.5.1
torchvision==0.14.1+cu116
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
zipp==3.15.0
